layout/style/test/test_mq_changes_in_iframe.html

<!DOCTYPE HTML>
<html>
<head>
  <meta charset="utf-8">
  <title>Media feature value change propagation in an iframe</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<p id="display"></p>
<div id="content" style="display: none"></div>
<iframe id="iframe"></iframe>
<pre id="test"></pre>
<script>
add_task(async () => {
  const mqString = "(prefers-reduced-motion: reduce)";

// Bug 1638618: Theme changes (and thus LookAndFeel, including reduced-motion)
// aren't propagated to processes that don't have PresContexts in them
// (i.e. only to processes with open tabs), and so preallocated processes
// have the Theme/LookAndFeel values current when they were prestarted.
// It can also affect e10s processes that are recycled via Provide/Take.
// Add an extra level of handshake to ensure we change reduced-motion
// after the iframe's process is loaded.   When that is fixed we could
// restore the pre-iframe-creation setting to false here.

  iframe.src = SimpleTest.getTestFileURL("mq_changes_child.html")
                         .replace("mochi.test:8888", "example.com");
  await new Promise(resolve => window.addEventListener("message", event => {
    if (event.data == "ready") {
      resolve();
    }
  }, { once: true } ));
  // Must be after loading the iframe (bug 1638618).
  // This may or make not trigger a 'change' event in the iframe, depending
  // on the initial value of reduced-motion the frame's process has.

  SpecialPowers.DOMWindowUtils.setPrefersReducedMotionOverrideForTest(false);

  const mql = matchMedia(mqString);
  // Wait if needed for the value to propagate
  if (mql.matches) {
    await new Promise(resolve => mql.addEventListener("change", resolve,
                                                      { once: true }));
  }
  ok(!mql.matches, `Doesn't match ${mqString}`);

  const changedInThisDocument = new Promise(resolve => {
    mql.addEventListener("change", event => { resolve(event.matches); });
  });
  const changedInIFrame = new Promise(resolve => {
    window.addEventListener("message", event => {
      if ("matches" in event.data) {
        resolve(event.data.matches);
      }
    }, { once: true });
  });

  // This will trigger a change event in the iframe always, and will send a
  // match event back to us
  SpecialPowers.DOMWindowUtils.setPrefersReducedMotionOverrideForTest(true);

  const results =
      await Promise.allSettled([ changedInThisDocument, changedInIFrame ]);

  results.forEach(result => {
    is(result.status, "fulfilled");
    ok(result.value, `Matches ${mqString}`);
  });

  SpecialPowers.DOMWindowUtils.resetPrefersReducedMotionOverrideForTest();
});
</script>
</body>
</html>
